[clients] Move Runtime
